Move code into src/ and update imports
